+++ README.md
@@ -4,48 +4,21 @@
 [![Test](https://github.com/devantler-tech/dotnet-helm-cli/actions/workflows/test.yaml/badge.svg)](https://github.com/devantler-tech/dotnet-helm-cli/actions/workflows/test.yaml)
 [![codecov](https://codecov.io/gh/devantler-tech/dotnet-helm-cli/graph/badge.svg?token=RhQPb4fE7z)](https://codecov.io/gh/devantler-tech/dotnet-helm-cli)
 
-<details>
-  <summary>Show/hide folder structure</summary>
-
-<!-- readme-tree start -->
-```
-.
-├── .github
-│   └── workflows
-├── scripts
-├── src
-│   └── Devantler.HelmCLI
-│       └── runtimes
-│           ├── linux-arm64
-│           │   └── native
-│           ├── linux-x64
-│           │   └── native
-│           ├── osx-arm64
-│           │   └── native
-│           ├── osx-x64
-│           │   └── native
-│           ├── win-arm64
-│           │   └── native
-│           └── win-x64
-│               └── native
-└── tests
-    └── Devantler.HelmCLI.Tests
-        └── HelmTests
-
-22 directories
-```
-<!-- readme-tree end -->
-
-</details>
-
 A simple .NET library that embeds the Helm CLI.
 
 ## 🚀 Getting Started
 
+### Prerequisites
+
+- .NET 9.0 or later
+- [Helm CLI](https://helm.sh/docs/intro/install/) installed and available in your system's PATH
+
+### Installation
+
 To get started, you can install the package from NuGet.
 
 ```bash
-dotnet add package Devantler.HelmCLI
+dotnet add package DevantlerTech.HelmCLI
 ```
 
 ## 📝 Usage
@@ -53,7 +26,7 @@ dotnet add package Devantler.HelmCLI
 You can execute the Helm CLI commands using the `Helm` class.
 
 ```csharp
-using Devantler.HelmCLI;
+using DevantlerTech.HelmCLI;
 
 var (exitCode, output) = await Helm.RunAsync(["arg1", "arg2"]);
 ```
