--- src/Devantler.HelmCLI/Devantler.HelmCLI.csproj
@@ -1,30 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net9.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <AnalysisMode>All</AnalysisMode>
-    <EnforceCodeStyleInBuild>true</EnforceCodeStyleInBuild>
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-    <GenerateDocumentationFile>true</GenerateDocumentationFile>
-    <CodeAnalysisTreatWarningsAsErrors>true</CodeAnalysisTreatWarningsAsErrors>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="CLIWrap" Version="3.9.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Content Include="runtimes/*/native/*">
-      <Link>%(Filename)%(Extension)</Link>
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-      <PackagePath>runtimes/</PackagePath>
-    </Content>
-  </ItemGroup>
-
-  <ItemGroup>
-    <InternalsVisibleTo Include="Devantler.HelmCLI.Tests" />
-  </ItemGroup>
-</Project>
